+++ packages/core/src/data/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './matchers';
 export * from './parsePath';
 export * from './postHandling';
-export * from './endpoints';
+export * from '../../utils/endpoints';