+++ packages/core/src/types.ts
@@ -36,4 +36,9 @@ export type HeadlessConfig = {
 	useWordPressPlugin?: boolean;
 	integrations?: Integrations;
 	sites?: HeadlessConfig[];
+	debug?: {
+		requests?: boolean;
+		redirects?: boolean;
+		devMode?: boolean;
+	};
 };