+++ packages/core/src/utils/getHeadlessConfig.ts
@@ -1,7 +1,13 @@
-import { endpoints } from '../data/utils/endpoints';
-import { HeadlessConfig } from '../types';
+import { endpoints } from './endpoints';
+import type { HeadlessConfig } from '../types';
 
-declare const __10up__HEADLESS_CONFIG: HeadlessConfig;
+let __10up__HEADLESS_CONFIG: HeadlessConfig = {};
+
+export function setHeadstartWPConfig(config: HeadlessConfig) {
+	__10up__HEADLESS_CONFIG = { ...config };
+}
+
+export const setHeadlessConfig = setHeadstartWPConfig;
 
 /**
  * Returns the contents of headless.config.js
@@ -13,7 +19,7 @@ declare const __10up__HEADLESS_CONFIG: HeadlessConfig;
  *
  * @returns The contents of headless.config.js
  */
-export function getHeadlessConfig() {
+export function getHeadstartWPConfig() {
 	const {
 		customPostTypes,
 		redirectStrategy,
@@ -23,6 +29,7 @@ export function getHeadlessConfig() {
 		sites,
 		hostUrl,
 		integrations,
+		debug,
 	} = __10up__HEADLESS_CONFIG;
 
 	const headlessConfig: HeadlessConfig = {
@@ -33,6 +40,7 @@ export function getHeadlessConfig() {
 		redirectStrategy: redirectStrategy || 'none',
 		useWordPressPlugin: useWordPressPlugin || false,
 		integrations,
+		debug,
 		sites: (sites || []).map((site) => {
 			// if host is not defined but hostUrl is, infer host from hostUrl
 			if (typeof site.host === 'undefined' && typeof site.hostUrl !== 'undefined') {
@@ -55,6 +63,7 @@ export function getHeadlessConfig() {
 	return headlessConfig;
 }
 
+export const getHeadlessConfig = getHeadstartWPConfig;
 /**
  * Get a config for a specific site
  *