+++ packages/next-redis-cache-provider/src/index.ts
@@ -13,10 +13,19 @@ export default class RedisCache implements CacheHandler {
 
 	private fs?: CacheFs;
 
+	/**
+	 * The build ID of the current build
+	 */
 	private BUILD_ID?: string;
 
+	/**
+	 * The server dist directory
+	 */
 	private serverDistDir?: string;
 
+	/**
+	 * The Redis client
+	 */
 	private redisClient: Redis;
 
 	constructor(ctx: CacheHandlerContext) {
@@ -26,10 +35,37 @@ export default class RedisCache implements CacheHandler {
 		this.redisClient = this.getRedisClient();
 	}
 
+	/**
+	 * Builds a Redis Client based on the environment variables
+	 *
+	 * If VIP_REDIS_PRIMARY and VIP_REDIS_PASSWORD are set, it will use those, otherwise
+	 * it will use NEXT_REDIS_URL
+	 *
+	 * @returns Redis client
+	 */
 	public getRedisClient() {
-		return new Redis((process.env.NEXT_REDIS_URL as string) ?? undefined);
+		const redisUrl = process.env.NEXT_REDIS_URL;
+
+		const endpoint = process.env.VIP_REDIS_PRIMARY;
+		const password = process.env.VIP_REDIS_PASSWORD;
+
+		if (endpoint && password) {
+			const [host, port] = endpoint.split(':');
+			return new Redis({
+				host,
+				password,
+				port: parseInt(port, 10),
+			});
+		}
+
+		return redisUrl ? new Redis(redisUrl) : new Redis();
 	}
 
+	/**
+	 * Gets an unique build id from the BUILD_ID file
+	 *
+	 * @returns The build ID from the BUILD_ID file
+	 */
 	public async getBuildId() {
 		if (!this.fs || !this.serverDistDir) {
 			return '';