+++ packages/next/jest.setup.ts
@@ -1,6 +1,6 @@
 import 'whatwg-fetch';
 
-import { server } from '@10up/headless-core/test';
+import { server } from '@headstartwp/core/test';
 
 // @ts-expect-error
 global.__10up__HEADLESS_CONFIG = {};