+++ projects/wp-multisite-nextjs/src/pages/_sites/[site]/index.js
@@ -1,4 +1,4 @@
-import { getHeadlessConfig } from '@10up/headless-core/utils';
+import { getHeadlessConfig } from '@headstartwp/core/utils';
 import {
 	usePost,
 	fetchHookData,
@@ -6,7 +6,7 @@ import {
 	addHookData,
 	handleError,
 	usePosts,
-} from '@10up/headless-next';
+} from '@headstartwp/next';
 import PropTypes from 'prop-types';
 import { PageContent } from '../../../components/PageContent';
 import { indexParams } from '../../../params';