+++ projects/wp-nextjs/src/pages/book/[...path].js
@@ -7,7 +7,7 @@ import {
 	handleError,
 	usePost,
 	useAppSettings,
-} from '@10up/headless-next';
+} from '@headstartwp/next';
 import { PageContent } from '../../components/PageContent';
 import { bookParams } from '../../params';
 import { resolveBatch } from '../../utils/promises';