+++ wp/headless-wp/plugin.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * Plugin Name: 10up Headless WordPress
+ * Plugin Name: HeadstartWP
  * Plugin URI:  false
  * Description: Adds functionality to the WordPress admin and REST API for 10up's headless framework.
- * Version:     0.6.0
+ * Version: 1.0.1
  * Author:      10up
  * Author URI:  https://10up.com
- * Text Domain: 10up-headless-wp
+ * Text Domain: headstartwp
  * Domain Path: /languages
  *
  * @package HeadlessWP
@@ -17,7 +17,7 @@ namespace HeadlessWP;
 use HeadlessWP\Preview\PreviewToken;
 
 // Useful global constants.
-define( 'HEADLESS_WP_PLUGIN_VERSION', '0.6.0' );
+define( 'HEADLESS_WP_PLUGIN_VERSION', '1.0.1' );
 define( 'HEADLESS_WP_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 define( 'HEADLESS_WP_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
 define( 'HEADLESS_WP_PLUGIN_INC', HEADLESS_WP_PLUGIN_PATH . 'includes/' );